hooks refactor
