+++ packages/web/src/pages/dashboard-page/components/constants.ts
@@ -1,1 +1,3 @@
 export const showMoreLimit = 5
+
+export const tablePageSize = 50
